// src/app/models/usa-booking.model.ts
export interface UsaBooking {
  id: number
  booking_number: string
  user_id: string
  sender_name: string
  sender_phone: string
  receiver_name: string
  receiver_phone: string
  receiver_address: string
  delivery_man_id: any
  assign_at: any
  road_at: any
  complete_at: any
  created_at: string
  updated_at: string
}